Update docstrings; fix utils.hpp
